Fix cleanup after early game load failures

+++ rts/Sim/Units/Scripts/UnitScriptEngine.cpp
@@ -52,6 +52,9 @@ void CUnitScriptEngine::InitStatic() {
 
 void CUnitScriptEngine::KillStatic() {
 	RECOIL_DETAILED_TRACY_ZONE;
+	if (unitScriptEngine == nullptr)
+		return;
+
 	cobEngine->Kill();
 	cobFileHandler->Kill();
 	unitScriptEngine->Kill();
